refactor(migration): move statehistory into state/history
