+++ migration/state/history/counter.go
@@ -1,4 +1,4 @@
-package statehistory
+package history
 
 import (
 	"time"
