+++ migration/state/history/parse.go
@@ -1,4 +1,4 @@
-package statehistory
+package history
 
 import (
 	"encoding/binary"
