+++ migration/state/history/storage_ingestor.go
@@ -1,4 +1,4 @@
-package statehistory
+package history
 
 import (
 	"bytes"
