+++ node/migration.go
@@ -14,8 +14,8 @@ import (
 	"github.com/NethermindEth/juno/migration/deprecated" //nolint:staticcheck,nolintlint,lll // ignore statick check package will be removed in future, nolinlint because main config does not check
 	"github.com/NethermindEth/juno/migration/historyprunner"
 	"github.com/NethermindEth/juno/migration/state/headstate"
+	"github.com/NethermindEth/juno/migration/state/history"
 	"github.com/NethermindEth/juno/migration/statedifflength"
-	"github.com/NethermindEth/juno/migration/statehistory"
 	"github.com/NethermindEth/juno/utils/log"
 )
 
@@ -32,7 +32,7 @@ func registerMigrations(cfg *Config) *migration.Registry {
 		).
 		WithOptional(&headstate.Migrator{}, cfg.NewState, "new-state").
 		With(&statedifflength.Migrator{}).
-		WithOptional(&statehistory.Migrator{}, cfg.NewState, "new-state")
+		WithOptional(&history.Migrator{}, cfg.NewState, "new-state")
 
 	return registry
 }
